ARM: imx: serial: Fixed a TXDC interrupts are infinite in RS485 mode
Data that does not fit in the Tx buffer (struct circ_buf) at once is
stored separately.

Problems can occur if data is stored in the Tx buffer between the time
the Tx DMA callback is called and the time the FIFO and Shift register
are both empty.

This is because the DMA callback is the DMAC BH(tasklet).

Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>
Change-Id: Id46ef48567c55ba6b4ceca1b0992c0947db9156a