serial: fsl_lpuart: handle break and make sysrq work
Although there is already sysrq characters handling, a break condition
was never detected. Add support for it.

The LPUART can't distinguish between a framing error and a break
condition. We assume it is a break if the received data is all zero.

Signed-off-by: Michael Walle <michael@walle.cc>
Link: https://lore.kernel.org/r/20210512141255.18277-6-michael@walle.cc
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 5541a9bacfe54fb2038531fb276724fcec6c29c4)
Change-Id: Id96d6bbe060995431d7e2011c4788126f14d04ea
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>