pca9450-regulator: add a restart handler through i2c
Handling system restart through PMIC instead of watchdog allows
distinguishing between a crash and a voluntary reboot by probing
the PWRON_STAT register on next boot.

Priority is set higher than psci's to be selected as default, and
failure to communicate with the pmic will fall back to watchdog
reset.

Note the i2c bus has likely been suspended and requires resuming,
this is done by keeping track of the i2c device and explicitely
calling pm_runtime_resume.
This is harmless if the device was not actually suspended, but
the function will then return an error so just continue after
warning.