net/wireless/nxp: add dts parameters per mac
the original nxp driver already had settings in fdt but these did not
allow setting different parameters for each mac handle, which is
something we want to do.

This defines a new top-level fdt block 'moal-params', which can either
take settings directly or in a subnode 'mac-X' where X is 0 or 1 for
each mac handle.

Example:

  moal-params {
    antcfg = <0x102>;
    mac-0 {
      drv_mode = <0x1>;
    };
  };

antcfg and drv_mode are the only two settings implemented but more can
easily be added if required.