imx8mp.dtsi: fix failed to reparent hsio_axi to sys_pll2_500m
The error as follows:

  clk: failed to reparent hsio_axi to sys_pll2_500m: -16

The parent of usb_root_clk is hsio_axi.

  hsio_axi
    usb_root_clk

In reset value of i.MX8M Plus, the parent of usb_root_clkhsio_axi is
sys_pll1_800m(It is different from the Reset value in the Reference
Manual).

  sys_pll1_800m
    hsio_axi

At this time, the parent of hsio_axi is reparent to sys_pll2_500m, but
it cannot be done. Cannot reparent if clock gate is open.

Therefore, set the parent of hsio_axi to sys_pll2_500m first.

  sys_pll2_500m
    hsio_axi
      usb_root_clk

Change-Id: Idfb22f218bf6f3745321ac69a68aa173e226d9dd
Fixes: e0b6018894b8 ("clk: imx8mp: fix usb_root_clk parent")
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>