net/wireless/nxp: add ps_mode/auto_ds dts params
We want to disable power savings by default, setting this from DTS
is easiest at this point.

Note we do not override the previous value of ps_mode/auto_ds if it was
non-zero (default): this allows forcing a value through modprobe
parameter, which should have priority over dts content.
This could not be done for e.g. drv_mode which does not have a default
zero value, making the dts value (if any) always win.

Also added the final value to /proc/mwlan/adapter0/config so it can be
possible to check what value is currently used, as it turns out the
module parameter is writable but writing into it has no effect...

Change-Id: Id7ef4e11b729ee828b7f8ba845a25bc9970352d2
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>