x2_defconfig: downgrade cfg80211 to module
having cfg80211 built-in would try to load regulatory.db before rootfs is
mounted, thus failing to load it.
Enabling it as module instead delays the loading until (and only if!) a
wireless card is detected, thus saving memory for all non-wireless
boards.

Since the wifi drivers already are enabled as module this should not
cause any extra problem either.

Change-Id: I69abbd7b44ee6fb2ac620bedf17a14b65db2ebc7
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>