arm64: armadillo_iotg_g4.dts: fix MDIO signal issue during suspend
This is currently only an issue with Armadillo boards with marvell
ethernet phys.  The issue is caused by the pull-up circuit added at
enet1_led3 signal line.  Since enet1_led3 is active low, it is drived
high during suspend-to-mem state and a MDIO signal is drived via the
pull-up circuit.

The workaround is setting pad settings of enet1_led3 to pull-down and
open-drain during suspend-to-mem state.

Change-Id: I9db6437a184b08533ccd1b86a8cb2abb24d437bd
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>