arm64: armadillo_iotg_g4.dts: make ethernet pins pull-down during suspend
During suspend-to-mem state, ethernet pins such as ENET_TD0 can not be
drived by anyone.  So these pins should be pulled down (or up).

A note about ENET_MDIO (for eqos) and ENET1_MDIO (for fec).  ENET_MDIO
must not be pulled down because it is pulled up on the board.  And in
contrast to ENET_MDIO, ENET1_MDIO must be pulled down because the
regulator for fec is disabled (pull-up is not effective) during
suspend-to-mem state.

Change-Id: I789031afcd11cd728057d3a116b412d2e1a3ed92
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>