x2_defconfig: add pcieuart9098_combo_v1 to EXTRA_FIRMWARE
Embed the nxp/pcieuart9098_combo_v1.bin firmware (832KB) into the
kernel.

the nxp driver is sensitive to firmware versions mismatch, so in order
to avoid problems with users who would upgrade the userland while
keeping an old kernel it is preferable to embed the firmware into the
kernel and not make it part of the userland.

Change-Id: I272643b71204d8ade61e86169eff08516fccf952
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>