arm: dts: armadillo-iotg-a6e-lwb5plus: Specify LWB5+ NVRAM firmware file name
LWB5+ NVRAM firmware is given a file name for each board. If the
firmware does not exist, fallback to "brcmfmac4373-sdio.txt".

If there is no firmware it will look for rootfs. Therefore, it takes
about 1 minute to fallback. This means that the WLAN cannot be used
for about 1 minute.

This is wasteful, so we use a fixed firmware file name.

Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>
Change-Id: I6c5e428bc6e35554091d57d056c1070bbb8568a9