arm: dts: armadillo-6x0: Fix ethernet phy is not reset
The reset property must be specified for mdio bus.

The reset timing may not be met, and correct phy access may not be
possible.

Reading BMSR regster before releasing reset may read 0xffff. Then the
initialization fails like this:
  SMSC LAN8710/LAN8720 2188000.ethernet-1:00: Master/Slave resolution failed, maybe conflicting manual settings?

This problem occurs when booting without the LAN cable connected.

The reason why the problem does not occur when the LAN cable is
connected is probably because the current flows from the current cable
and resets faster.

Change-Id: Ie4edec0024b0323563e64c69d23ff12faea587e6
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>