clk: fractional-divider: Improve approximation when zero based and export
Consider the CLK_FRAC_DIVIDER_ZERO_BASED flag when finding the best
approximation for m and n. By doing so, increase the range of valid
values for the numerator and denominator by 1.

Furthermore, export the approximation function so that users of this
function can be compiled as modules.

Cc: A.s. Dong <aisheng.dong@nxp.com>
Signed-off-by: Frank Oltmanns <frank@oltmanns.dev>
Link: https://lore.kernel.org/r/20230617131041.18313-2-frank@oltmanns.dev
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit 2e9abc6ec108370ad341f98d0e05e02e4e165b5a)
[ In the case of hardware such as i.MX 8ULP, where the division value
is 0-based (e.g., 7 is specified in the field when dividing by 8),
even if the frequency at which the division value for the parent clock
is maximized is set to assigned-clock-rates, the frequency actually
generated is one step below the maximum.
For example, in the case of i.MX 8ULP, a maximum of 7 can be specified
in the PCD field of the PCC register, which corresponds to dividing
the parent clock by 8. Therefore, when the frequency of the parent
clock is 192 MHz and 24 MHz is specified in the assigned-clock-rate,
7 is expected to be written to the PCD field. However, the program
actually writes 6 (27.428571 MHz) because the zero base is not taken
into account when calculating the value of the PCD field.
This commit solve this problem.
Unlike the situation in the original commit,
the clk_fd_general_approximation function was static, so the
EXPORT_SYMBOL_GPL line was removed to suppress the warning output. ]
Change-Id: I2098aa833e3cce5364267356a74acea13da9380c
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>