LF-10915-1 gpio: gpio-imx-rpmsg: use common irq call back to handle the real rpmsg operation
[ IMX 6.6 tree commit ab052fa11cc66f30118e8a18075584af703fd7e8 ]

For the mask and unmask, need in irq contex, so can't handle rpmsg
operation directlly(it may sleep and schedule), the previous method
use the workqueue, which will trigger the following issue:

BUG: scheduling while atomic
[  339.548280][    T8]  schedule+0x84/0xf0
[  339.548294][    T8]  schedule_timeout+0x90/0x12c
[  339.548316][    T8]  wait_for_common+0xdc/0x1a8
[  339.548331][    T8]  wait_for_completion_timeout+0x14/0x24
[  339.548347][    T8]  mbox_send_message+0xc4/0x158
[  339.548377][    T8]  imx_rproc_kick+0x40/0xac [imx_rproc]
[  339.548424][    T8]  rproc_virtio_notify+0x40/0x54
[  339.548446][    T8]  virtqueue_kick+0x104/0x13c
[  339.548467][    T8]  rpmsg_send_offchannel_raw+0x380/0x438 [virtio_rpmsg_bus]
[  339.548512][    T8]  virtio_rpmsg_send+0x2c/0x3c [virtio_rpmsg_bus]
[  339.548554][    T8]  rpmsg_send+0x38/0x5c
[  339.548567][    T8]  gpio_send_message+0x64/0x14c [gpio_imx_rpmsg]
[  339.548609][    T8]  imx_rpmsg_irq_set_wake+0x90/0x67c [gpio_imx_rpmsg]
[  339.548647][    T8]  irq_set_irq_wake+0xac/0x1d8
[  339.548669][    T8]  max17042_suspend+0x34/0x4c [max17042_battery]

And the irq common code did contain call_backs(bus lock,
bus_sync_unlock) for slow bus, and can avoid upper issue.
So switch to irq_bus_lock/irq_bus_sync_unlock.

Currently the gpio set and get share the pin's gpio_rpmsg_data buffer.
This is not safe when multi user try to handle the same gpio pin.

Take this into consideration. Process A try to set the gpio value.
after config the nesessary information into the gpio_rpmsg_data, CPU
schedule to process B. In the same time, Process B try to read the value
from the same gpio pad. and sucessfull get the return value into the
pin's gpio_rpmsg_data. Then CPU schedule to Process A, and continue do
the following thing, but this time, the gpio_rpmsg_data are fulshed by
porcess B, not the original value.

Though we use mutex lock in gpio_send_message(), but this is not safe
enough. Need to involve the lock at the every beginning of the gpio
call back function.

Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Change-Id: Ia884a239676cb38b4e8c1093ee33fcda62ad3f92
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>