rpmsg_ns: fix initialization of driver_override
driver_override is freed when stopping the remoteproc as follow:
# echo stop > /sys/class/remoteproc/remoteproc0/state

so driver_override should be an allocated string and not a static value.

This was fixed in the following two commits upstream:
42cd402b8fd4 ("rpmsg: Fix kfree() of static memory on setting driver_override")
bb17d110cbf2 ("rpmsg: Fix calling device_lock() on non-initialized device")

which have already been backported to 5.10, but rpmsg_ns is not part of
upstream 5.10 so this part of the commits had been dropped.

Change-Id: Ife1e9773b8b8782c71c909f9486c04a40375b4c8
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>