phy-fsl-samsung-hdmi: add pixel clock for 32.0/51.2/65.0/71.0MHz
Add pixel clock to use displays which do not support any standard HDMI
frequencies.

The frequencies are based on the previous closest value with a small
divider updated to provide a rough approximation, further work will be
done to further tune the values and/or automatically select something
approximate for other non-standard displays.

Change-Id: I201fc2bd8ebd08b7088b13ddacad3611fe32fde5
Signed-off-by: Makoto Sato <makoto.sato@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>