arch/arm/armadillo*dts: add explicit alias for rtc1 as well
When linux boots, it will assign rtc0 to the first clock to be set up if
it has no alias set, so if the snvs rtc happens to be registered first
it'd take the rtc0 slot and the external rtc would fall back to rtc1
with just a message in dmesg.

Thanksfully the bringup depends only on how the DTB was built and it
turns out the external RTC added later to the DTB is probed first, but
let's not leave that up to chance and make sure snvs would get assigned
rtc1 even if it happens to be probed early.

Change-Id: Ief487a185fb4a1df5804b1daa8dd0aee2ea03c36
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>