gpio-imx-rpmsg: move back to workqueue
The new bus lock code does not work when the irq is not threaded or does
not have a callback, e.g. something like gpiomon will setup a thread that
locks/unlocks the bus on every ack but something like gpio-keys will not,
in which case the first interrupt will be processed and nothing after it
ever will.

Using a work-queue does not have this problem, because we can queue work
in irq context as everything is pre-allocated.

The original work-queue code had problems because it tried to share a
single work item between all GPIOs, just pre-allocate one item per
pin.
The logic has been kept closer to the new code, but probably will need
to be reworked later as this is definitely not correct in some sequence

Change-Id: Ia54ded7e553c2244c93b33565d188fc6f2a680d8
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>