arm64: dts: armadillo_iotg_g4: disable pull enable for SD card
Disabling the pull-up in the resistor settings allows some SD cards to
work properly. The reason is that when pull-up is enabled, it is
outside the range of the SD standard.

These signal lines are pulled up to 15kOhm on the board.  If the
pull-up is enabled in the resistor settings, the minimum resistance at
3.3V is 18kOhm, connected in parallel with 15kOhm.  The resistance
value of this combined resistor is approximately 8kOhm, which is
outside the range of the SD standard.

Redmine: 65185
Change-Id: I61b0f5dc9ec381d9de7f56eaf1a710f94cca3a3f
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>