arm64: dts: armadillo_iotg_a9: use runtime-pm as the default for usdhc2
If you want the SD card to always be powered on, apply
armadillo_iotg_a9-usdhc2-alwayson.dtbo.

When you connect an SD card, USDHC2_CD goes low, Linux detects the SD
card is connected, and power is supplied to the SD card. When
USDHC2_CD goes high, Linux detects the SD card is disconnected, and
power is turned off for the SD card.

Change-Id: I21a89da2bc80c6ec9636c647884484a018aea276