i2c: imx: prevent rescheduling in non dma mode
We are experiencing a problem with the i.MX I2C controller when
communicating with SMBus devices. We are seeing devices time-out because
the time between sending/receiving two bytes is too long, and the SMBus
device returns to the idle state. This happens because the i.MX I2C
controller sends and receives byte by byte. When a byte is sent or
received, we get an interrupt and can send or receive the next byte.

The current implementation sends a byte and then waits for an event
generated by the interrupt subroutine. After the event is received, the
next byte is sent and we wait again. This waiting allows the scheduler
to reschedule other tasks, with the disadvantage that we may not send
the next byte for a long time because the send task is not immediately
scheduled. For example, if the rescheduling takes more than 25ms, this
can cause SMBus devices to timeout and communication to fail.

This patch changes the behavior so that we do not reschedule the
send/receive task, but instead send or receive the next byte in the
interrupt subroutine. This prevents rescheduling and drastically reduces
the time between sending/receiving bytes. The cost in the interrupt
subroutine is relatively small, we check what state we are in and then
send/receive the next byte. Before we had to call wake_up, which is even
less expensive. However, we also had to do some scheduling, which
increased the overall cost compared to the new solution. The wake_up
function to wake up the send/receive task is now only called when an
error occurs or when the transfer is complete.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>
Acked-by: Oleksij Rempel <o.rempel@pengutronix.de>
Signed-off-by: Andi Shyti <andi.shyti@kernel.org>
(cherry picked from commit 5f5c2d4579ca6836f5604cca979debd68ecfe23f)
[ We experienced similar problems with ADS1115, especially under high
load situations such as kernel startup. Due to this problem, when
the Config register of ADS1115 is rewritten with regmap_update_bits(),
incorrect data may be written to the Config register. This leads to
fictitious interrupts from the ADS1115.
When cherry-picking, we ignored the difference inherent in slave mode
because it is not an expected use case. ]
Change-Id: I8ac2cd42e91b81f6ba4500da0228acd035cd8543
Redmine: 67212
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>