nvmem: imx-ocotp-fsb-s400: fix fuse write offset
comparing ele_write_fuse() with ahab_write_fuse() in uboot, the index
argument must be the "word index" (e.g. bank * 8 + word) and not the
offset in the file in byte, so we need to divide by 4 (word size)

Ensure offset is a multiple of 4 and fix index value properly.

Fixes: "LF-11560-2 nvmem: imx-ocotp-fsb: support program fuse for i.MX93/95"
Change-Id: Ice7a78ef301e9e498ebb66a462b077e8586e2378
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>