nvmem: imx-ocotp-fsb-s400: add sysfs file for secureboot
secureboot lifecycle is managed outside of fuses on i.MX8ULP:
the SRK hash itself is stored in fuse, but close state is not.

Instead, close state can be checked through a register of the fsb,
and modified through the ele firmware.

Since checking requires access to the fsb mapping we implement
access in this driver, through an extra sysfs file:
- reading the file will show open/close and any violation events
- writing 'close' will close the device if it is open

Note closing does not check for violation events!
This is because if we install a known good image we can write
SRK and close immediately without rebooting, meaning there will
be violations at the time

Change-Id: Iba7fde5406ef67380164a64bf4ad57ed775c72bf
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>