a6*_defconfig: enable W1 modules, HWMON
Enable HWMON (built-in) and one-wire (W1, modules) in armadillo a6
defconfig, as per debian configuration.

For ABOS we want to have commonly pluggable devices available by default
without recompiling a kernel.

This comes at a size increase of 12KB for uImage (for hwmon) and 224KB
for the w1 modules, or 50KB compressed.

Link: https://armadillo.atmark-techno.com/forum/armadillo/23128
Change-Id: I27800bc56b30f37014ef0820f1fb227b4f9b0c4f
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>