reset: sim7672: fix power not stay on, at DL mode
If USB_BOOT pin is set to low and the power is turned on to goto F/W
update mode, regulator is turned off and the process ends without
status pin becoming running.

To perform F/W update mode, if USB_BOOT pin is low, the process ends
without checking the status.

Change-Id: I0141317cca62675c174bcafd6f76c42b03627466
Signed-off-by: Mitsuhiro Yoshida <mitsuhiro.yoshida@atmark-techno.com>