serial: imx_rpmsg: hardcode name to ttyrpmsg0
changing the rtos side to add or remove endpoints will change rpdev->dst,
but we need the linux name to be fixed.

We can add an ida later if we start supporting more than one channel,
but the current protocol can only ever support one so just leave as is
with an hardcoded number.

Change-Id: I7994c53c0091f3902b5291bf69a39f97138d57d4
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>