mtd: spi-nor: atmel: Add support for AT25QL128A serial NOR flash
In the upstream kernel, the generic SPI-NOR driver parses the SFDP
information to detect flash chips. However, this functionality is missing
in the 5.10-cip kernel. To enable detection of the AT25QL128A serial NOR
flash on 5.10-cip, this patch adds an entry for the chip in the driver.

Link for the AT25QL128A serial NOR flash datasheet:
https://www.renesas.com/en/document/dst/at25ql128a-datasheet?r=1608586

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Message-ID: <20241206090524.66639-2-prabhakar.mahadev-lad.rj@bp.renesas.com>
Signed-off-by: Pavel Machek <pavel@denx.de>
Signed-off-by: Nobuhiro Iwamatsu <iwamatsu@nigauri.org>