serial: imx_rpmsg: set the initial baud rate
tty_std_termios was the default setting, so it was set to 38400bps.
 # stty -F /dev/ttyrpmsg0 speed
 38400

However, the default setting in the m33 firmware is 115200bps. Fixed
the Linux device driver and m33 firmware to have the same baud rate.

This also fixes the issue where UART data cannot be received
immediately after booting. When this issue occurs, the following
message is displayed on the m33 console:
 no channel, never got anything from linux?

Change-Id: I3f1a34187923072d55d67ac487f2cac3c20048a0
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>