arm64: dts: a9e: rename dts that can be shared with a900
nousb and usdhc2-alwayson only require minor changes to nodes in
armadillo_900.dts, so they can always be used on armadillo_900.

armadillo_iotg_a9e-sim7672 uses functions not included in the
recommended design for armadillo-900 modules (gpiof 25, gpiod 16,
etc.), so it may not be usable if the motherboard is anything other
than the armadillo-iot gateway a9e.

Change-Id: I1750a7d159cfa6f13c31dc1eec2536029038617d
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>