i2c: rpmsg-imx: do not access received data after cb return
The cb data is recycled immediately after the cb returns, so we must not
just store 'msg' in the global i2c_rpmsg for the xfer thread to access
later.

Instead, copy values we're interested in:
- always copy ret_val
- if a buffer was requested to copy, and the length matches, then copy
  the data directly in cb.

This does the same amount of work so should not even be slower, the copy
just happens in a different, safe, context.

Change-Id: If246e87355e79ad8e26f88980d1563854757d034
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>