pwm: imx-rpchip: set rpdev last in rpmsg probe
The platform driver probe can try to use the mutex/completion as soon as
rpdev is set in theory, so set the rpdev last.

(This might be missing a barrier too if a race really happens here? For
now this is more a theorical problem than anything else, so this will
do.)

Change-Id: I5b7214a467f96ede1761821fd1810e76adf23b44
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>