arm64: dts: a9e: assign label to VIN iio device
Assigns a label to the iio device corresponding to the ADC that measures
the voltage value of VIN.

The value defined as label can be referenced from the label file under
the iio:device[x] directory. This is useful for specifying iio device
and is also referenced by gw_app and power-alertd.

Change-Id: I189b995236cfede899b3f9c4e91174ff1d10daa8
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>