iio: adc: imx-rpmsg: add the ability to read in_voltage_scale
To obtain a voltage value using an ADC, it is necessary to know the
voltage value per 1 LSB as well as the raw register value that stores
a ADC result.

Generally, the voltage value per 1 LSB is calculated from the reference
voltage and resolution,
In the ADC of i.MX 8ULP, in addition to this, it is necessary to
consider that the input voltage is divided by the value of CSCALE.

When CSCALE is 1, no voltage divider is applied and the voltage is at
full scale, but when CSCALE is 0, the input voltage is divided by 30/64.

With this commit, the in_voltage_scale can be read out, which also
takes into account the value of CSCALE described in the device tree.

Change-Id: Iae33801a104c392e94a145b8dd71a538971e9712
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>