arm64: dts: a9e: change PTB12/14 to open drain
PTB12/14 corresponds to DI(Digital Input) 1/2 respectively on a9e.
PTB12/14 is intended to be used as Input, it can also be configured as
Output, and push-pull output is available by default.

Since PTB12/14 is connected to the collector terminal of the
photocoupler, if the push-pull output PTB12/14 goes High when DI1/2 is
High by an external signal, a shoot-through current will flow.

Set PTB12/14 to open drain so that no shoot-through current flows even
if PTB12/14 goes High.

Incidentally, tried the following two other solutions, but they did
not solve this problem.

- Set IBE (input buffer enable) instead of ODE.
  -> PTB12/14 can be set as output.

- Set GPIO hog definition as input in a child node of the GPIO
  controller.
  -> can not control PTB12/14 using gpioget/set command.

Change-Id: I1eba431f78f6d0c12d0cdd20fdd419be571c1e0b
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>