drivers: imx-mailbox: allow disabling wakeup in dtb
When receiving a message after the suspend callback, the imx-mailbox
driver triggers a wakeup but on i.MX8ULP linux cannot safely abort
suspend anymore at this point.

The problem i.MX8ULP has is that drivers/soc/imx/rpmsg_life_cycle.c
expects a reply but both drivers suspend callbacks are processed in
parallel, so it's possible for the lifecycle reply to come in after
mailbox's and abort suspend.
It's possible to correctly abort suspend in m33 side by adding yet
another timeout, but in this particular case we do not want to abort
suspend, and this is simpler than adding proper ordering between the
drivers as an immediate solution.

Change-Id: I703c86a84f302dec01eadf1aec7476a444a4855c
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>