gpio: imx-rpmsg: fix initial wakeup bug
The phenomenon that wakeup by the pin specified by setwake could not be
performed at the first suspend after power-on or reboot occurred with a
high probability (suspend after the second time was no problem).

The cause of the bug is that the wakeup set(=1) by enable_irq_wake() is
overwritten with 0 by imx_rpmsg_gpio_send_ack() which is called in
gpiod_to_irq() called for the first time in suspend_fn().

gpiod_to_irq() behaves differently the first time and the second and
later times: the first time it looks for a mapping for irq, but since
there is none, it creates a new mapping; the second and later times,
since there is already a mapping created, it ends at the stage of
looking for a mapping for irq, and never creates a new mapping.

The process of creating a mapping for irq calls imx_rpmsg_irq_ack(), and
it set queue_work that cause imx_rpmsg_gpio_send_ack to be executed with
a time delay. This is the function that disables wakeup(=0).

To solve this bug, take quick fix approach that call gpiod_to_irq() in
the setwake phase instead of suspend_fn().

Change-Id: If87704a30eee1b443eb104a80585830f297161b6
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>