arm64: dts: a9e: disable PTB2 DFE and PTB2 PE
When DFE of PTB2 was enabled, input events could not be obtained, so
DFE was removed.  There was no chattering with the tactile switch
currently installed on the a9e, so there should be no problem even
without DFE.

Also, since PTB2 is pulled up on the a9e board, PE was removed.

Change-Id: I52d49650125617604d840790e9f21d0d9ebec583
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>