arm64: dts: a900, a9e: move description of USB0 regulator from armadillo_900.dts
armadillo_900.dts contained a description of the USB0 regulator, but
originally the Armadillo-900 did not have a specification for USB VBUS.

Therefore, that information, which is specific to the A9E and
Armadillo-900 evaluation boards, was moved to the respective dts.

armadillo_900-nousb.dts is no longer a minor modification of
armadillo_900.dts and has been renamed armadillo_iotg_a9e.

Change-Id: Ib0a2ce33481ccbc97db44d0812fcc17854489e1f
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>