drm/bridge: nwl-dsi: fix MIPI lane range
The normal range for lane count is 1 to 4, and this is the case for
other lines as well.

This line was the only one that had a range of 2 to 4 for no reason,
so I determined this was incorrect and fixed it. Operation of a
display that uses only one lane was also confirmed.

Change-Id: I79fd0fb5639146bb9c183e1be8f777566b3cd1d5
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>