localversion: remove -cip*, add -at2
When a user builds their kernel from tarball they only get the version
from the Makefile, making it impossible to distinguish between e.g. our
versions 5.10.236-r1 from 5.10.236-r0

In order to track this add a new localversion-at that we'll reset to 0
on every stable merge, and increment when releasing r1+ subpackages.
This commit will be r2 release so initialize to at2.

Since the cip version is redundant remove the cip file, it will be
removed on cip merges

Change-Id: Id1acedaf4d93c202de2208b535dccae391d5c53a
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>